requirements.txt
requests
fpdf2
google-genai
pytest
python-dotenv
